+++ CHANGELOG.md
@@ -16,9 +16,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python version `3.10` added to package classifiers
 - `should_ask` keyword argument for `Echo` and `Acknowledge` interactions ([#356](https://github.com/wayfair-incubator/columbo/issues/356))
 
+### Changed
+
+- `BasicQuestion.ask()` will only evaluate dynamic values for the prompt message and default value once instead of
+  repeatedly when the response was invalid.
+
+### Fixed
+
+- Prevent infinite loop when the default value for a `BasicQuestion` does not satisfy the `Validator` and `no_user_input`
+  was set to `True`. Now raises a `ValueError` when this situation is detected.
+
+  The intent was that the default value would always satisfy the `Validator`, but that was not enforced or explicitly
+  documented.
+
 ### Removed
 
-* Support for Python version `3.6`.
+- Support for Python version `3.6`.
 
 ## [0.11.0] - 2021-08-04